Rename project to simple-air-comfort
