+++ .github/ISSUE_TEMPLATE/bug-report.yml
@@ -6,7 +6,7 @@ body:
     attributes:
       value: |
         Thanks for taking the time to fill out this bug report!
-        If you just need help, please use the [ring-tile discusson](https://community.home-assistant.io/t/ring-tile-card-visualise-your-sensor-data/899257) on the community forums.
+        If you just need help, please use the [simple-air-comfort discusson](https://community.home-assistant.io/t/simple-air-comfort-card-visualise-your-sensor-data/899257) on the community forums.
         **NOTE!** I may take me a while to get back to you. Please bear with me!
   - type: textarea
     id: ha_version
@@ -52,9 +52,9 @@ body:
   - type: textarea
     id: minimal_code
     attributes:
-      label: Include the YAML code you are using with the ring-tile card causing trouble
+      label: Include the YAML code you are using with the simple-air-comfort card causing trouble
       placeholder: |
-        type: custom:ring-tile
+        type: custom:simple-air-comfort
         entity: sensor.temperature
       render: YAML
   - type: textarea
