+++ .github/workflows/build-release.yml
@@ -16,23 +16,23 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Pre check
-        run: ls -al /home/runner/work/ring-tile-card/ring-tile-card/
+        run: ls -al /home/runner/work/simple-air-comfort-card/simple-air-comfort-card/
 
       - name: Build
         run: |
-          cd /home/runner/work/ring-tile-card/ring-tile-card
+          cd /home/runner/work/simple-air-comfort-card/simple-air-comfort-card
           npm install
           npm run build
 
       - name: Post check
-        run: ls -al /home/runner/work/ring-tile-card/ring-tile-card/dist/
+        run: ls -al /home/runner/work/simple-air-comfort-card/simple-air-comfort-card/dist/
 
         # Upload build file to the release as an asset.
       - name: Upload as release asset
         uses: svenstaro/upload-release-action@v2
         with:
           repo_token: ${{ secrets.GITHUB_TOKEN }}
-          file: /home/runner/work/ring-tile-card/ring-tile-card/dist/ring-tile-card.js
-          asset_name: ring-tile-card.js
+          file: /home/runner/work/simple-air-comfort-card/simple-air-comfort-card/dist/simple-air-comfort-card.js
+          asset_name: simple-air-comfort-card.js
           tag: ${{ github.ref }}
           overwrite: true
