+++ LICENCE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2025 ring-tile-card for Home Assistant
+Copyright (c) 2025 simple-air-comfort-card for Home Assistant
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
