+++ hacs.json
@@ -1,5 +1,5 @@
 {
-  "name": "Ring Tile Card",
+  "name": "Simple Air Comfort Card",
   "homeassistant": "2025.5.0",
   "render_readme": true
 }
