+++ src/simple-air-comfort-card.js
@@ -1,4 +1,4 @@
-import { RingTile } from "./ring-tile";
+import { RingTile } from "./simple-air-comfort";
 import { addFonts } from "./helpers/ringTileFonts";
 import { RtInfo } from "./rt-info";
 import { RtRing } from "./rt-ring";
@@ -6,22 +6,22 @@ import { RtRingSvg } from "./rt-ring-svg";
 import * as pkg from "../package.json";
 
 console.info(
-  `%c ring-tile-card %c v${pkg.version} `,
+  `%c simple-air-comfort-card %c v${pkg.version} `,
   'color: yellow; font-weight: bold; background: darkblue',
   'color: white; font-weight: bold; background: dimgray',
 );
 
 addFonts();
 
-customElements.define("ring-tile", RingTile);
+customElements.define("simple-air-comfort", RingTile);
 customElements.define("rt-info", RtInfo);
 customElements.define("rt-ring", RtRing);
 customElements.define("rt-ring-svg", RtRingSvg);
 
 window.customCards = window.customCards || [];
 window.customCards.push({
-  type: "ring-tile",
-  name: "Ring Tile Card",
+  type: "simple-air-comfort",
+  name: "Simple Air Comfort Card",
   preview: true,
   description:
     "Add a ring to your sensor tile cards to visualise sensor state.",
